+++ CHANGES.md
@@ -5,6 +5,8 @@
 - Fix Docker compose file, so example data can be loaded into database (author @zstatmanweil, https://github.com/stac-utils/stac-fastapi-pgstac/pull/142)
 - Handle `next` and `dev` tokens now returned as links from pgstac>=0.9.0
 
+- Fix `filter` extension implementation in `CoreCrudClient`
+
 ## [3.0.0] - 2024-08-02
 
 - Enable filter extension for `GET /items` requests and add `Queryables` links in `/collections` and `/collections/{collection_id}` responses ([#89](https://github.com/stac-utils/stac-fastapi-pgstac/pull/89))